// src/types.ts
export class RouteMeta {
  title: string;

  constructor({ title }: { title: string }) {
    this.title = title;
  }
}